Break down Search component
