+++ client/src/components/Header/Header.css
@@ -1,0 +1,20 @@
+.header {
+  background-color: #0077b5;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  min-height: 50px;
+  padding: 0.5rem 1rem;
+  color: #fff;
+  margin-bottom: 1rem;
+}
+
+.logo {
+  width: 200px;
+  user-select: none;
+}
+
+.searchbox {
+  width: 600px;
+  margin-bottom: 3rem;
+}
