+++ client/src/components/ResultList/ResultList.css
@@ -1,0 +1,14 @@
+.search-panel__results {
+  width: 100%;
+  flex-direction: column;
+  margin: 30px;
+}
+
+.pagination {
+  margin: 2rem auto;
+  text-align: center;
+}
+
+ul {
+  justify-content: center;
+}
